requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
pillow==10.1.0
